Add Last-Write-Wins functionality
If a timestamp is provided in the patch options, changes will be optionally applied if they are the last write, otherwise they will ignored. Timestamps are stored in a property named `$lww$` on the object.

Only "add", "remove", and "replace" operations are permitted with LWW. It is meant for basic operations only.