Add Last-Write-Wins functionality
Adding a utilities that will help sync an object field-by-field using last-writer-wins. This sync method is limited but simple. It does not work with array items, though entire arrays can be set. It doesn't work with "move" or "copy" operations, only "add", "remove", and "replace" operations are permitted with LWW. It should work great for documents like Figma describes in https://www.figma.com/blog/how-figmas-multiplayer-technology-works/ and for objects like user preferences.

It works by using a metadata object to track the current revision of the object, any outstanding changes needing to be sent to the server, and the revisions of each added value so that one may get all changes since a given revision. It will be very small on the client, and only moderately sized on the server. It is up to the implementor to store this metadata object with the rest of the data. These are tools to help you deal with the harder part of LWW syncing, but you'll have to implement them in your system.

It should work with clients offline, though those clients will "win" when they come back online and write to the server. If this is not desired, simply send the data from the server down when first connecting and then just receive changes.

It includes a whitelist or blacklist (not both) of properties that cannot be set by the client, only set by the server itself.