Provide a more sure way to check if a path is an array
We can only really know if a path is an array vs a property on an object by providing the source object the ops are meant to be applied to. In order to prevent a backwards incompatible change, I made it an optional parameter at the end of `rebase` but it really _should_ be used in every case. We could remove the ? and make it required in the future.